Now icons on buttons draw in its center
