+++ Gravity_Simulation/UI/Button.py
@@ -19,12 +19,6 @@ class Button(UIObject):
 
         # FIXME: сделать так, чтобы иконка была по центру
 
-        new_icon_surface = pygame.Surface((self._width, self._height))
-        rect = self._icon.get_rect()
-        rect.move(width // 2 - rect.width // 2, height//2 - rect.height // 2)
-        new_icon_surface.blit(self._icon, rect)
-        self._icon = new_icon_surface
-
         self._action = action
 
     def event_handler(self, event):
@@ -39,16 +33,6 @@ class Button(UIObject):
 
         self.state_handler()
 
-    def is_mouse_in_rect(self):
-        mouse_pos = mouse.get_pos()
-        return self.rect.collidepoint(*mouse_pos)
-
-    def change_hover_or_focus_state(self):
-        if self.is_mouse_in_rect():
-            self._state = MouseState.HOVER
-        else:
-            self._state = MouseState.FREE
-
     def state_handler(self):
         if self._state == MouseState.HOVER:
             self.on_mouse_hover()
@@ -57,7 +41,9 @@ class Button(UIObject):
 
     def draw(self, surface: Surface):
         pygame.draw.rect(surface, self._color, self.rect, border_radius=4)
-        surface.blit(self._icon, self.rect)
+        icon_rect = self._icon.get_rect()
+        icon_rect = icon_rect.move(self._x + self._width // 2 - icon_rect.width // 2, self._y + self._height // 2 - icon_rect.height // 2)
+        surface.blit(self._icon, icon_rect)
 
     def on_mouse_hover(self):
         self._color = (190, 190, 190)
