+++ Gravity_Simulation/UI/UIObject.py
@@ -6,6 +6,7 @@ from abc import ABC
 from abc import abstractmethod
 from pygame.rect import Rect
 from enum import Enum
+import pygame.mouse
 
 
 class MouseState(Enum):
@@ -54,6 +55,16 @@ class UIObject(ABC):
     @abstractmethod
     def on_mouse_leave(self):
         pass
+
+    def is_mouse_in_rect(self):
+        mouse_pos = pygame.mouse.get_pos()
+        return self.rect.collidepoint(*mouse_pos)
+
+    def change_hover_or_focus_state(self):
+        if self.is_mouse_in_rect():
+            self._state = MouseState.HOVER
+        else:
+            self._state = MouseState.FREE
 
     @abstractmethod
     def on_mouse_click(self):
